+++ src/__generated__/MapRendererQuery.graphql.ts
@@ -31,6 +31,10 @@ query MapRendererQuery(
 fragment GlobalMap_viewer_3La17j on Viewer {
   city(slug: $citySlug) {
     name
+    sponsoredContent {
+      introText
+      artGuideUrl
+    }
     coordinates {
       lat
       lng
@@ -245,7 +249,7 @@ return {
   "kind": "Request",
   "operationKind": "query",
   "name": "MapRendererQuery",
-  "id": "94c718c869e5629d247bde4591173de0",
+  "id": "58706a5aed1c0e765b9ebbfb053ae7ba",
   "text": null,
   "metadata": {},
   "fragment": {
@@ -317,6 +321,31 @@ return {
             "plural": false,
             "selections": [
               v1,
+              {
+                "kind": "LinkedField",
+                "alias": null,
+                "name": "sponsoredContent",
+                "storageKey": null,
+                "args": null,
+                "concreteType": "CitySponsoredContent",
+                "plural": false,
+                "selections": [
+                  {
+                    "kind": "ScalarField",
+                    "alias": null,
+                    "name": "introText",
+                    "args": null,
+                    "storageKey": null
+                  },
+                  {
+                    "kind": "ScalarField",
+                    "alias": null,
+                    "name": "artGuideUrl",
+                    "args": null,
+                    "storageKey": null
+                  }
+                ]
+              },
               v2,
               {
                 "kind": "LinkedField",